Function & Type Documentation

Index

BCIInterface.DataHandler

BCIInterface.DataProcessor

BCIInterface.Device

BCIInterface.Experiment

BCIInterface.FFTDataDescriptor

BCIInterface.MCP3208

BCIInterface.ModelData

BCIInterface.RawDataDescriptor

BCIInterface.StandardProcessor

BCIInterface.Standard

BCIInterface._read_metadata

BCIInterface._write_metadata

BCIInterface.clear!

BCIInterface.clear!

BCIInterface.create_data

BCIInterface.create_data

BCIInterface.create_data

BCIInterface.createpath

BCIInterface.gather_data!

BCIInterface.get_sample!

BCIInterface.get_voltage

BCIInterface.iscompatible

BCIInterface.process

Documentation

BCIInterface.DataHandler — Method

DataHandler(data_processor::DataProcessor, data_io::DataIO; cases=nothing,
name=nothing, max_freq=nothing)

Create a DataHandler instance. cases, name and max_freq are automatically determined by data saved at path if they are nothing.

Example:

data_io = DataIO("data/test", states)
data_handler = DataHandler(data_io, StandardFFT())

BCIInterface.DataProcessor — Type

Abstract class containing data processors.

Defined data processors: StandardProcessor: The standard processors with preset arguments and functions, for details see StandardProcessor.

BCIInterface.Device — Type

Device for gathering EEG data. Create it using

Device(board::EEGBoard)

... TODO

BCIInterface.Experiment — Method

Experiment(device::Device, name::String; tags::Array=[], 
extra_info::Dict=Dict(), path::String="data/", load_previous::Bool=false)

name: Name of the experiment (e.g. "BlinkDetection").

tags: Tags which will be applied to all data gathered with this Experiment.

path: To top-level of data directory (e.g. "data/"). If empty, files can't be saved.

TODO: descs for keywords TODO: load_previous not implemented yet (maybe in another function?)

BCIInterface.FFTDataDescriptor — Method

FFTDataDescriptor(num_channels::Int, max_freq::Int)

Data descriptor for FFT data in the format of a 2D-Array per sample, with the first dimension corresponding to the frequncies from 1 to maxfreq and the seconds dimension corresponding to the channels from 1 to numchannels.

BCIInterface.MCP3208 — Method

MCP3208(path::String, num_channels::Int; max_speed_hz::Int=1000, 
online=true)

Initialise a MCP3208-based self-built EEG.

path::String: SPI path of the device

num_channels::Int: Number of connected electrodes.

max_speed_hz::Int: 1000 by default. Tells the Raspberry Pi (or similar device) how often per second the output should be read.

TODO: explain (side) effects

online::Bool: true by default; if set to false, get_voltage(...) will just return 0. This is for testing all associated functions and devices using this board without having it connected (e.g. in tests).

BCIInterface.ModelData — Method

ModelData(data::Data, outputs::Dict; batchsize=5, shuffle=true)

Create ModelData suitable for use with neural networks.

data:

tag2output:

BCIInterface.RawDataDescriptor — Method

RawDataDescriptor(num_channels::Int)

Data descriptor for raw data in the format of a 1D-Array per sample, with each channels voltage in a row.

num_channels::Int: Number of channels the data has (= width of the row).

BCIInterface.StandardProcessor — Type

Standard configuration for processing EEG data. It uses a preset of functions and options and may not work for you.

Create with Standard.

BCIInterface.Standard — Method

Standard()::StandardProcessor

Create standard configuration for processing EEG data. See StandardProcessor for more details.

BCIInterface._read_metadata — Method

_read_metadata(filepath)

Internal method used for reading from file; in function to make later switch of file format easier.

BCIInterface._write_metadata — Method

_write_metadata(filepath, metadata::Metadata)

Internal method used for writing to file; in function to make later switch of file format easier.

BCIInterface.clear! — Method

clear!(data::Data)

Delete all saved data.

BCIInterface.clear! — Method

clear!(experiment::Experiment)

Delete all saved raw data from experiment.

BCIInterface.create_data — Method

create_data(name::String, data_desc::RawDataDescriptor)

Create new Data-Object for FFT data.

BCIInterface.create_data — Method

create_data(name::String, data_desc::RawDataDescriptor)

Create new Data-Object for raw data (MCP3208).

BCIInterface.create_data — Method

create_data(name::String, device::Device)

Create Data-Object which fits given device (raw data, fft data, etc.).

BCIInterface.createpath — Method

create_path(path::String)

Create necessary folders and file if they don't exist yet, so that isdir(path) or isfile(path) returns true depending on whether path points to a folder or a file.

BCIInterface.gather_data! — Method

gather_data!(experiment::Experiment, runtime::Number; tags::Array=[], 
extra_info::Dict=Dict())

Gather raw EEG data.

runtime: Runtime in seconds.

Optional arguments:

tags: Tags to add to every data point on top of tags given to Experiment

extra_info: Extra info to add to every data point on top of extra info given to Experiment

BCIInterface.get_sample! — Method

get_sample!(board::EEGBoard)

Updates board.sample to new data from board.

BCIInterface.get_voltage — Method

get_voltage(board::MCP3208, channel::Int)

Read digital output of the ADC on channel channel.

If MCP3208 was given online=false at creation, return 0 instead.

BCIInterface.iscompatible — Method

is_compat(df::DataFrame, metadata::Union{Metadata, Nothing})

Check if df is "compatible" with metadata.

If metadata is nothing, then return true (for easy use with load_metadata which returns nothing when no metadata is defined)

BCIInterface.process — Method

process(device::Device, processor::StandardProcessor)

TODO: Processing

1. Artifacts Removal
2. Data Filtering
3. Feature Extraction (FFT)

Also see this, page 2529.