Gathering, processing and loading EEG data

Introduction

Gathering data

To gather EEG data, you will first need a physical EEG device. We are currently developing our own which you can build yourself at home with the instrucations in The MCP3208 – a cheap alternative for commercial EEGs.

This framework currently only offers direct support for this EEG. However, support can be easily extended, for instructions see Using your own EEG device.

The following steps assume that you already have the hardware set up. If you have added your own EEG, just replace MCP3208 with the custom function you created.

To gather EEG data, you need to create a Device object with Device(board::EEGBoard). An Example:

device = Device(MCP3208("/dev/spidev0.0", 8))

For details about MCP3208, see MCP3208.

Processing data

Reference

BCIInterface.EEG.Channel

BCIInterface.EEG.DataProcessor

BCIInterface.EEG.Device

BCIInterface.EEG.MCP3208

BCIInterface.EEG.StandardProcessor

BCIInterface.EEG.Standard

BCIInterface.EEG.process

BCIInterface.EEG.Channel — Type

Single Channel with voltages in Volt and the times the voltages were recorded at in seconds.

BCIInterface.EEG.DataProcessor — Type

Abstract class containing data processors. StandardProcessor: The standard processors with preset arguments and functions.

BCIInterface.EEG.Device — Type

Device for gathering EEG data. Create it using

Device(board::EEGBoard)

... TODO

BCIInterface.EEG.MCP3208 — Method

MCP3208(path::String, num_channels::Int; max_speed_hz::Int=1000, online=true)

Initialise a MCP3208-based self-built EEG. path refers to the SPI path of the device and num_channels to the number of connected electrodes.

With online set to true, the EEG will be "simulated": get_voltage(...) will return random values.

TODO: maxspeedhz

BCIInterface.EEG.StandardProcessor — Type

Standard()::StandardProcessor

For more details, see

Standard()

BCIInterface.EEG.Standard — Method

Standard()::StandardProcessor

Create standard configuration for processing EEG data. It uses a preset of functions and options and may not work for you.

BCIInterface.EEG.process — Method

TODO: Processing

1. Artifacts Removal
2. Data Filtering
3. Feature Extraction (FFT)

Also see this, page 2529.