Function & Type Documentation

Index

BCIInterface.AbstractDataDescriptor

BCIInterface.AbstractDataProcessor

BCIInterface.Data

BCIInterface.Data

BCIInterface.Data

BCIInterface.DataHandler

BCIInterface.EEGBoard

BCIInterface.Experiment

BCIInterface.FFTDataDescriptor

BCIInterface.MCP3208

BCIInterface.Model

BCIInterface.Model

BCIInterface.ModelData

BCIInterface.ModelData

BCIInterface.RawDataDescriptor

BCIInterface.StandardProcessor

BCIInterface.Standard

BCIInterface._read_metadata

BCIInterface._write_metadata

BCIInterface.clear!

BCIInterface.clear!

BCIInterface.create_model

BCIInterface.createpath

BCIInterface.datetime

BCIInterface.element

BCIInterface.gather_data!

BCIInterface.get_most_recent

BCIInterface.get_sample!

BCIInterface.get_timestr

BCIInterface.get_voltage

BCIInterface.iscompatible

BCIInterface.load_model

BCIInterface.parse_modelname

BCIInterface.process

BCIInterface.save

BCIInterface.save

BCIInterface.save

BCIInterface.save

BCIInterface.standard_network

BCIInterface.standard_trainingparameters

BCIInterface.testmode!

BCIInterface.train!

BCIInterface.trainmode!

Documentation

BCIInterface.AbstractDataDescriptor — Type

Abstract type containing data descriptors.

Data is processed and trained on differently depending on the data descriptor. If you store or use your data in a format which doesn't have a data descriptor yet, you will need to define your own and overload the following functions:

create_data(name::String, data_desc::MyDataDescriptor)

get_inputshape(data_desc::MyDataDescriptor)

inputarray(data_descriptor::MyDataDescriptor, num_samples::Int)

row2inputs(data_descriptor::MyDataDescriptor, row::Vector)

BCIInterface.AbstractDataProcessor — Type

Abstract class containing data processors.

Defined data processors:

StandardProcessor: The standard processors with preset arguments and functions, for details see StandardProcessor.

BCIInterface.Data — Method

Data(name::String, board::EEGBoard; dir="data/")

Create Data-Object which fits given board (raw data, fft data, etc.).

BCIInterface.Data — Method

Data(name::String, data_desc::FFTDataDescriptor; dir="data/")

Create new Data-Object for FFT data.

BCIInterface.Data — Method

Data(name::String, data_desc::RawDataDescriptor; dir = "data/")

Create new Data-Object for raw data (MCP3208).

BCIInterface.DataHandler — Method

DataHandler(data_processor::DataProcessor, data_io::DataIO; cases=nothing,
name=nothing, max_freq=nothing)

Create a DataHandler instance. cases, name and max_freq are automatically determined by data saved at path if they are nothing.

Example:

data_io = DataIO("data/test", states)
data_handler = DataHandler(data_io, StandardFFT())

BCIInterface.EEGBoard — Type

Supertpye of all EEG boards.

When creating your own EEG board, you should make it a subtype of AbstractEEGBoard with

struct MyBoard <: AbstractEEGBoard
    ...
end

BCIInterface.Experiment — Method

Experiment(board::EEGBoard, name::String; tags::Array=[],
extra_info::Dict= Dict(), path::String="data/", load_previous::Bool=false)

name: Name of the experiment (e.g. "BlinkDetection").

tags: Tags which will be applied to all data gathered with this Experiment.

path: To top-level of data directory (e.g. "data/"). If empty, files can't be saved.

TODO: descs for keywords

BCIInterface.FFTDataDescriptor — Method

FFTDataDescriptor(num_channels::Int, max_freq::Int)

Data descriptor for FFT data in the format of a 2D-Array per sample, with the first dimension corresponding to the frequncies from 1 to max_freq and the seconds dimension corresponding to the channels from 1 to num_channels.

BCIInterface.MCP3208 — Method

MCP3208(path::String, num_channels::Int; max_speed_hz::Int=1000,
online=true)

Initialise a MCP3208-based self-built EEG.

path::String: SPI path of the device

num_channels::Int: Number of connected electrodes.

max_speed_hz::Int: 1000 by default. Tells the Raspberry Pi (or similar device) how often per second the output should be read.

TODO: explain (side) effects

online::Bool: true by default; if set to false, get_voltage(...) will just return 0. This is for testing all associated functions and devices using this board without having it connected (e.g. in tests).

BCIInterface.Model — Type

Model(network::Flux.Chain, epochs::Int, savedir::String, modelname::String)

Create a Model object with four fields:

network: The neural network in the form of a Flux Chain.

epochs: The number of epochs the model has already been trained for.

savedir: The directory where the model will be saved.

modelname: The name of the subdirectory where the model will be saved.

BCIInterface.Model — Method

(model::Model)(inputs)

Apply model to inputs to calculate outputs.

Example:

model(inputs)

BCIInterface.ModelData — Type

ModelData

A type for storing data in a neural network friendly way.

Fields:

dataloader::DataLoader: Store the data in a Flux.DataLoader. Data can be accessed with dataloader.data, returning a named tuple (inputs=..., outputs=...). Both fields of the tuple are multidimensional arrays with the last dimension corresponding to the sample.

tag2output: see outputs described in the constructor ModelData.

datadescriptor: The AbstractDataDescriptor of the original data.

BCIInterface.ModelData — Method

ModelData(data::Data, outputs::Dict; batchsize=5, shuffle=true)

Create ModelData suitable for use with neural networks.

data: A [Data] object storing the EEG data.

tag2output: Dictionary; each key is a tag (saved in data) with the value representing the output the model should give for that tag.

Example:

# classification:
data_class = load_data("data/Emotions", "RawData")
tag2output_class = Dict(
    :scared => [1.0, 0.0, 0.0],
    :angry => [0.0, 1.0, 0.0],
    :happy => [0.0, 0.0, 1.0]
)
modeldata_class = ModelData(data, tag2output_class)
# regression:
data_reg = load_data("data/Concentration", "RawData")
tag2output_reg = Dict(
    :rest => [0.0],
    :task1 => [0.2],
    :task2 => [0.7],
    :task3 => [1.0]
)
modeldata_reg = ModelData(data, tag2output_reg)

BCIInterface.RawDataDescriptor — Method

RawDataDescriptor(num_channels::Int)

Data descriptor for raw data in the format of a 1D-Array per sample, with each channels voltage in a row.

num_channels::Int: Number of channels the data has (= width of the row).

BCIInterface.StandardProcessor — Type

Standard configuration for processing EEG data. It uses a preset of functions and options and may not work for you.

Create with Standard.

BCIInterface.Standard — Method

Standard()::StandardProcessor

Create standard configuration for processing EEG data. See StandardProcessor for more details.

BCIInterface._read_metadata — Method

_read_metadata(filepath)

Internal method used for reading from file; in function to make later switch of file format easier.

BCIInterface._write_metadata — Method

_write_metadata(filepath, metadata::Metadata)

Internal method used for writing to file; in function to make later switch of file format easier.

BCIInterface.clear! — Method

clear!(data::Data)

Delete all saved data.

BCIInterface.clear! — Method

clear!(experiment::Experiment)

Delete all saved data.

BCIInterface.create_model — Method

create_model(data::ModelData; network_constructor::Function=standard_network)

Create a model with a neural network for classifying / regressing data.

network_constructor: The given function is given three arguments: the input shape and output shape of the data, each represented by a tuple, and the data descriptor of type DataDescriptor.

The default network_constructor if not supplied is standard_network.

BCIInterface.createpath — Method

create_path(path::String)

Create necessary folders and file if they don't exist yet, so that isdir(path) or isfile(path) returns true depending on whether path points to a folder or a file.

BCIInterface.datetime — Method

datetime(seconds_since_epoch)

Convert seconds since unix epoch into a DateTime object.

BCIInterface.element — Method

element(dict::Dict)

Return the value of the first key-value pair of dict.

BCIInterface.gather_data! — Method

gather_data!(experiment::Experiment, runtime::Number; tags::Array=[],
extra_info::Dict=Dict())

Gather raw EEG data.

runtime: Runtime in seconds.

Optional arguments:

tags: Tags to add to every data point on top of tags given to Experiment

extra_info: Extra info to add to every data point on top of extra info given to Experiment

BCIInterface.get_most_recent — Method

get_most_recent(dir_path::String)::String

Return the path of the file with the most recent date in the file name. Only works on files where the data starts at the seventh character and has the format YYYY-mm-dd_HH-MM-SS. ````

BCIInterface.get_sample! — Method

get_sample!(board::EEGBoard)

Updates board.sample to new data from board.

BCIInterface.get_timestr — Method

get_timestr()::String

Return the current time as a string of the form YYYY-mm-dd_HH-MM-SS.

Used for file and folder names.

BCIInterface.get_voltage — Method

get_voltage(board::MCP3208, channel::Int)

Read digital output of the ADC on channel channel.

If MCP3208 was given online=false at creation, return 0 instead.

BCIInterface.iscompatible — Method

is_compat(df::DataFrame, metadata::Union{Metadata, Nothing})

Check if df is "compatible" with metadata.

If metadata is nothing, then return true (for easy use with load_metadata which returns nothing when no metadata is defined)

BCIInterface.load_model — Method

load_model(path::String)::Model

Load the model saved in file at path.

If path is a directory, use get_most_recent to find file with the most recent time in the file name.

BCIInterface.parse_modelname — Method

parse_modelname(modelname::String)

Parse all special characters in given string.

Special characters:

*t -> replace with current date and time (see get_timestr).

Example:

julia> parse_modelname("mymodel_*t_blue")
"mymodel_2023-02-04_14-17-51_blue"

BCIInterface.process — Method

process(board::EEGBoard, processor::StandardProcessor)

TODO: Processing

1. Artifacts Removal
2. Data Filtering
3. Feature Extraction (FFT)

Also see this, page 2529.

BCIInterface.save — Function

save(model::Model, path::String, overwrite=false)

Save given model at path.

If path is a directory, save the model there as a file with the current time as filename.

If overwrite=false (default), throw an error if the file at path already exists.

BCIInterface.save — Method

save(data::Data, df::DataFrame; name="data", checkmeta=true, updatemeta=true,
    overwrite=false)

Save given df. Folder is determined by data (data.dir and data.name), file name by name.

Keywords

checkmeta=true: Load and check Metadata saved at the location for compatibility if it already exists.

updatemeta=true: Save the updated Metadata.

overwrite=false: When true, overwrite already existing data and meta data.

BCIInterface.save — Method

save(data::Data; checkmeta=true, updatemeta=true, overwrite=false, repeat=false)

Save given data. Folder is determined by data (data.dir * "/" * data.name), file name by name.

BCIInterface.save — Method

save(model::Model)

Save the model in the folder model.savedir, in the subfolder model.modelname with the filename "modelcurrentTime", where currentTime is the current time as given by [`gettimestr`](@ref).

BCIInterface.standard_network — Method

standard_network(inputshape, outputshape, datadescriptor)::Flux.Chain

Choose and create a neural network based on the given inputshape, outputsshape and datadescriptor.

These networks are what we're currently using and may not be ideal for you.

There are only networks predefined for RawDataDescriptor and FFTDataDescriptor with a 1d output.

BCIInterface.standard_trainingparameters — Method

standard_trainingparameters()

Returns a TrainingParameters object with all fields populated with (hopefully) reasonable defaults.

BCIInterface.testmode! — Method

trainmode!(model::Model)

Some layers of neural networks behave differently when training and when testing the network (e.g. Dropout).

Use this to set the model into testing mode (Dropout etc. will not be executed.)

Use trainmode! for putting the model into training mode.

BCIInterface.train! — Method

train!(model::Model, data::ModelData;
params::TrainingParameters=standard_trainingparameters(), kws...)

Train the given model using given data as training data.

To adjust training parameters, you can either create your own TrainingParameters object and pass it as params=myparams or adjust the default TrainingParameters object (given by standard_trainingparameters) by passing key-value pairs like epochs=3, η=0.02, autosave=3.

Examples:

train!(model, modeldata, epochs = 20, show_progress = false, opt_rule = GradientDescent)

myparams = TrainingParameters(0.02, 20, x -> x, cpu, Flux.logitcrossentropy,
        GradientDescent, [], false, 5)

train!(model, modeldata, params = myparams)

BCIInterface.trainmode! — Method

trainmode!(model::Model)

Some layers of neural networks behave differently when training and when testing the network (e.g. Dropout).

Use this to set the model into training mode (Dropout etc. will be executed.)

Use testmode! for putting the model into testing mode.

Base.:== — Method

Base.:(==)(a::Flux.Chain, b::Flux.Chain)

Check two Flux Chains for equality. Returns true if

number and types of layers are the same

all fields (weight, bias, etc.) of all layers are the same

Base.:== — Method

Base.:(==)(a::Model, b::Model)

Check two Models for equality.

Example:

a::Model
b::Model

if a == b
    println("both models are the same!")
end