Gathering, processing and loading EEG data

Introduction

Gathering data

To gather EEG data, you will first need a physical EEG device. We are currently developing our own which you can build yourself at home with the instrucations in The MCP3208 – a cheap alternative for commercial EEGs.

This framework currently only offers direct support for this EEG. However, support can be easily extended, for instructions see Using your own EEG device.

The following steps assume that you already have the hardware set up. If you have added your own EEG, just replace MCP3208 with the custom function you created.

To gather EEG data, you need to create a Device object with Device(board::EEGBoard). An Example:

device = Device(MCP3208("/dev/spidev0.0", 8))

For details about MCP3208, see MCP3208.

Processing data

Reference

BCIInterface.EEG.Channel

BCIInterface.EEG.Data

BCIInterface.EEG.DataHandler

BCIInterface.EEG.DataProcessor

BCIInterface.EEG.Device

BCIInterface.EEG.Experiment

BCIInterface.EEG.MCP3208

BCIInterface.EEG.StandardProcessor

BCIInterface.EEG.Standard

BCIInterface.EEG._read_metadata

BCIInterface.EEG._write_metadata

BCIInterface.EEG.clear!

BCIInterface.EEG.createpath

BCIInterface.EEG.gather_data!

BCIInterface.EEG.iscompatible

BCIInterface.EEG.process

BCIInterface.EEG.Channel — Type

Single Channel with voltages in Volt and the times the voltages were recorded at in seconds.

BCIInterface.EEG.Data — Method

Create new Data-Object for raw data.

BCIInterface.EEG.DataHandler — Method

DataHandler(data_processor::DataProcessor, data_io::DataIO; cases=nothing, name=nothing, max_freq=nothing)

Create a DataHandler instance. cases, name and max_freq are automatically determined by data saved at path if they are nothing.

Example:

data_io = DataIO("data/test", states)
data_handler = DataHandler(data_io, StandardFFT())

BCIInterface.EEG.DataProcessor — Type

Abstract class containing data processors.

Defined data processors: StandardProcessor: The standard processors with preset arguments and functions, for details see StandardProcessor.

BCIInterface.EEG.Device — Type

Device for gathering EEG data. Create it using

Device(board::EEGBoard)

... TODO

BCIInterface.EEG.Experiment — Method

name: Name of the experiment (e.g. "BlinkDetection").

tags: Tags which will be applied to all data gathered with this Experiment.

path: To top-level of data directory (e.g. "data/"). If empty, files can't be saved.

TODO: load_previous not implemented yet (maybe in another function?)

BCIInterface.EEG.MCP3208 — Method

MCP3208(path::String, num_channels::Int; max_speed_hz::Int=1000, online=true)

Initialise a MCP3208-based self-built EEG. path refers to the SPI path of the device and num_channels to the number of connected electrodes.

With online set to true, the EEG will be "simulated": get_voltage(...) will return random values.

TODO: maxspeedhz

BCIInterface.EEG.StandardProcessor — Type

Standard configuration for processing EEG data. It uses a preset of functions and options and may not work for you.

Create with Standard.

BCIInterface.EEG.Standard — Method

Standard()::StandardProcessor

Create standard configuration for processing EEG data. See StandardProcessor for more details.

BCIInterface.EEG._read_metadata — Method

Internal method used for reading from file; in function to make later switch of file format easier.

BCIInterface.EEG._write_metadata — Method

Internal method used for writing to file; in function to make later switch of file format easier.

BCIInterface.EEG.clear! — Method

clear!(experiment::Experiment)

Delete all data from experiment.

BCIInterface.EEG.createpath — Method

create_path(path::String)

Create necessary folders and file if they don't exist yet, so that isdir(path) or isfile(path) returns true depending on whether path points to a folder or a file.

BCIInterface.EEG.gather_data! — Method

gather_data!(experiment::Experiment, runtime::Number; tags::Array=[], extra_info::Dict=Dict())

Gather raw EEG data.

runtime: Runtime in seconds.

Optional arguments:

tags: Tags to add to every data point on top of tags given to Experiment

extra_info: Extra info to add to every data point on top of extra info given to Experiment

TODO: create test

BCIInterface.EEG.iscompatible — Method

is_compat(df::DataFrame, metadata::Union{Metadata, Nothing})

Check if df is "compatible" with metadata.

If metadata is nothing, then return true (for easy use with load_metadata which returns nothing when no metadata is defined)

BCIInterface.EEG.process — Method

TODO: Processing

1. Artifacts Removal
2. Data Filtering
3. Feature Extraction (FFT)

Also see this, page 2529.